forms save to separate JSON files, appends to existing data or creates file and new object array, forms clear out upon submit and alert notifies client records are saved
