+++ data/equipData.json
@@ -1,14 +1,15 @@
 {
   "equipment": "Backhoe",
-  "date": "2025-01-13T00:00:00.000Z",
-  "type": "Preventative",
-  "technician": "DJ",
+  "date": "2025-01-08T00:00:00.000Z",
+  "type": "Repair",
+  "technician": "Rosie The Riveter ",
   "hours": 3,
-  "description": "replace air filter",
+  "description": "broken radiator hose ",
   "partsReplaced": [
-    "air filter",
-    ""
+    {
+      "part": "radiator hose"
+    }
   ],
-  "priority": "Medium",
+  "priority": "High",
   "completion": "Complete"
 }
