+++ data/maintenanceData.json
@@ -1,0 +1,50 @@
+[
+  {
+    "equipment": "Backhoe",
+    "date": "2025-01-13T00:00:00.000Z",
+    "type": "Repair",
+    "technician": "Rob Bot",
+    "hours": 3,
+    "description": "replace cracked oil pan",
+    "partsReplaced": [
+      {
+        "part": "oil pan"
+      }
+    ],
+    "priority": "High",
+    "completion": "Complete"
+  },
+  {
+    "equipment": "Excavator",
+    "date": "2025-01-01T00:00:00.000Z",
+    "type": "Preventative",
+    "technician": "Arnold",
+    "hours": 2,
+    "description": "oil change",
+    "partsReplaced": [
+      {
+        "part": ""
+      }
+    ],
+    "priority": "Low",
+    "completion": "Complete"
+  },
+  {
+    "equipment": "Backhoe",
+    "date": "2025-01-04T00:00:00.000Z",
+    "type": "Emergency",
+    "technician": "Rob Bot",
+    "hours": 5,
+    "description": "smoke from engine",
+    "partsReplaced": [
+      {
+        "part": "engine"
+      },
+      {
+        "part": "oil pan"
+      }
+    ],
+    "priority": "High",
+    "completion": "Complete"
+  }
+]
